Remove erroneous count statement
This item fixes issue #1808 by removing the useless and offending count
statement. The rest of the code should be revised however since it would
seem that the Validator assumed that an object should be in argument
but the ValueObject indicates a string.

This fix is beyond the scope of this issue and will be addressed separately.