Allow providing multiple templates using a comma-separated list
The `--template` option used to allow multiple templates to be invoked by
providing a comma-separated list. Apparently this broke during changed in
the past few releases.

In this change I have added a piece of code that will re-interpret the
discovered list of templates and explode them when a comma is found. To be sure
I also trim the output in case people manage to sneak in a space.