Show 'File' for globals function in Clean template
The Clean Template shows a File field when working with a global function but
the wrong variable names were used to display and link to the path of the
file.

In this change I have replaced the erroneous variable names with those that
should be used and as such restore the file relation.