Markers should not fatal when the marker contains special characters
In issue #1842 it was reported that the following marker causes failures:

    // @TODO This is a todo

This was caused by the @ in the marker as the Xml writer tries to use that
in the name of an element, which is not allowed per the XML standards.

We have changed the writer to strip all non-alphanumeric characters from the
name which makes the string safe to be used as an element name.