Mock statistics class
Since the getDestination method returns an invalid character when
running on windows for vfsstream we use a mock to override the actual
method. This method should not have any impact on the tested class.
We need to figure out a way to check the working of this method. Or move it
to the transform class.