Make sure $ sign is present in fqsen for properties
This change tackles an issue where the fqsen in a propertyDescriptor was missing a $ sign.
The fqsen would be something like this: \Namespace\Classname::property instead of \Namespace\Classname::$property.
The $ sign was added in the ElementsIndexBuilder, but that didn't fix the fqsen which is being used to display the property name in a @see or @link tag. With this commit the $ sign is added in the correct place.

Also the test coverage of PropertyDescriptor is completed to 100%