Add namespace to descriptors
The inline see tags in docblocks of properties and methods could not be resolved for elements in the same namespace as the container class because the property and method descriptors were missing their namespace. This way the context was not complete for the see tags and the relevant element could not be found. With this change the property and method assemblers will set the namespace in the descriptors.