Fix @see links in responsive template
Regular @see links were not visible as text and not working as links in the responsive template.
The links were treated the same way ase the license, link and author tags, but because they may reference structural elements, this doesn't always work. With this PR the @see and @uses tags are dealt with in a separate template. And when there is no 'http' in the link, they will be processed using the already existing phpDocumentor\Plugin\Core\Xslt\Extension::path