Adjust the xslt extension
The xslt extension has been extended to include a method with which you can discover the type of element you are dealing with.
Based on the outcome, you can take a different approach to building a link in the template.
The responsive template uses this method to figure out if it has to append the root to the path in case of a documented element, if it has to simply return the given link in case of a url, or if it has to return a # so the link will do nothing.