Make sure classes that are represented in namespace aliases get the correct reference
Relative classes that were represented in a namespace alias were treated as ambiguous type and passed on to the Linker for further investigation.
This way they got the wrong namespace, namely the namespace of the container class, and they couldn't be found in the element list.
With this commit a check is added to see if an element is represented by a namespace alias.

Also urls with scheme is ftp:// are left unchanged.