Removes dependency on cilex
Cilex is blocking the updates of symfony components. This
pr removes a lot of magic from the commands which were fetching dependencies
form the application.
The application is updated with some logic to access the inner container to support
the old behavior of the application like cilex provides. This needs to be removed
in the future.