Replace Cilex with Symfony4
In this change I have replaced Cilex with Symfony 4. In this change I have
introduced a compatibility layer so that third-party's can integrate with
minimal effort.

An important thing to note is that this layer should be temporary and there
is not a 100% guarantue. For one; order now matters and this may cause errors
for some third-parties.