Reintroduce processing markers from source code
Our previous reflection component was able to find markers in the source code.
We removed this behaviour during refactoring to focus on reflecting the source tree and
not it's contents.
After some benchmarking I found that this would be the most suitable way of implementing the
marker feature in phpdocumentor it self.