Refactor away the PathResolverFactory
I wanted to write a test for the PathResolverFactory and found out that it was
better to move the template directory defining functionality into the app class
a directly pass that method call into the PathResolver's arguments.

It would be nice to have this path as a config option or parameter but for now
this is an improvement that would reduce complexity.