Use cache folder in config for Twig
Prior to this change the Twig cache files were stored in the system's temp
directory. Although this works in many cases, it may lead to issues when
different projects have different twig files.

This change includes two things:

1. Change pipelines to read stages using tags
2. Introduce a new stage that sets the Twig cache folder to use a subfolder
   of the configured cache folder.

By changing the pipelines we can dynamically introduce a new stage in the
pipeline with a specific priority. This new stage reads the configuration
and preconfigures the twig service.