Simplify configuration
In this change I have taken the first steps to simplify configation management
in phpDocumentor.

With this change I have move the Configuration into an object and that object
in turn into the container. The ConfigurationFactory now can create the default
configuration from either the phpdoc.dist.xml _or_ the phpdoc.xml
(reintroduced).

By moving the configuration into the container we can do various things:

1. Inject the configuration into a service so that it can get to the settings
2. Use Configurators to inject configuration values into depending services
   (see https://symfony.com/doc/current/service_container/configurators.html)
3. Change the configuration and have those changes perpetuated in all locations
   where it is used. This makes it easier to adjust the configuration based on
   command line arguments or a custom defined configuration file.

An additional side effect of moving the configuration into the container is
that caching in the factory is no longer needed, and that various pipeline
stages can be changed into configurators.