Omitting the command name should execute run _with_ options and arguments
In this change I have changed the way the Symfony Console application
determines the active command name. Because we want users to be able to omit
run or project:run we check the first segment of the arguments if this is a
supported command.

If it isn't, we assume that it will be project:run and all options and
arguments are passed to that.