Inject the locations of the default configuration files
In a prior version the locations of the default configuration files were
hardcoded into the ConfigurationFactory. This caused a hard dependency on
the actual file system and prevented unit tests from touching every part
of the code.

By moving the definition of the default configuration files into the container
and injecting these values into the factory this part is abstracted away and
files may be provided from any source, including stubbing / mocking in a unit
test.

During development it became clear that Symfony does not expose the current
working directory by default, and since we need to fetch the default files
from that location a method was added in the Kernel class so that we can
use that as an expression in the service container.

Initially I tried to define it in the initializeContainer step in the Kernel,
but this does not work since that method only gets called when the cache is
stale and the container needs to be rebuilt. If the container cache is fresh
the old value will be re-used. This means that if you run phpDocumentor for
multiple projects, that it will only remember the working directory for when
it built the cache.