Restore orginal pre v3 type system
In the new phpdocumentor we changed the usage of type arrays to new
reflection oriented types. This caused a number of bc issues in templates.
To fix this I switched back to use the array notation.

The setTypes and getTypes methods of descriptors are now deprecated. And
should be replaced by the non pural variants.

refs #2020
fixes #1974