Fix regresion on phpdoc v2 config parsing
phpdoc v2 was parsing all methods by default. The new config
processor was not using this default. Beside that it was not
able to process multiple visibility definitions like it did before.

A remaining issue is that phpdoc v3 has a different definition for the
visibilty. We need to check how this can be compatible with eachother.

refs #2058
fixes #2056