Reintroduce the 'none' configuration file option
For testing purposes it is needed to be able to disable reading of the
default configuration files. This can also help users who have a default
file but do nto want to use that; for example a hosting party that wants
complete control over the configuration and disallow the files in the
default locations.