Refactor away the DebugEvent
The Debug Event was only used by the transformer and that module can just as
easily call the logger directly. I did keep the log and debug methods of the
transformer class as these could be used by Writers to log output so that they
need not know about phpDocumentor's logging mechanisms, and for backwards
compatibility.