Remove translation support from phpDocumentor
Translations were barely used as a feature and introduce extra complexity to
the codebase. In the future we may re-visit this decision but for now it
simplifies the codebase.

(also: in the future we may want to restrict ourselves to Symfony translations
as these integrate well with the rest of the framework)