Remove basic plugin support
During a Slack conversation Jaap and I agreed to remove plugin support for
phpDocumentor v3 pending a re-implementation.

This action will simplify the code and remove the need to make compatibility
layers. This is the first change to make this happen and in subsequent commits
I will dismantle all Plugins and rewire them into the container.

This will, in future commits, also remove any need for the Cilex compatibility
layer so that we can remove it.