Refactor DomainModel folder away since it is a partial implementation
At one point we wanted to adopt a more strict DDD model by also dividing the
application into Application, Infrastructure and Domain Model. This effort
was not rolled out completely and now the application was stuck in two
different architectures.

As an effort to simplify the application I have rolled back the Domain Model
folder back into the Domain Context folders as it was previously.