Make inclusion of sourcecode configurable
During the transition from phpDocumentor 2 to 3 we changed the way
configuration is loaded. This caused the sourcecode flag to become unusable.

Unfortunately, with phpDocumentor 3 option flags can only alter existing
configuration options and the sourcecode flag was never a configiration option.

With this change the sourcecode flag now overwrites the 'include-source'
configuration option. Int he process I have also adjusted the clean template
to only show the source button when the flag is set and adjusted the error
message to make it clear you cannot use that feature on localhost.

In the process I tightened up a couple of actions; such as forcing a rebuild
of the files when the visibility or sourcecode option changes between builds
and there have been a few bugs, for one with markers, that became obvious
and have been fixed in the process.