Add additional check in the Class Descriptor
Apparently there are sometimes errors that relate to the Property tag
collection containing non-property elements. Although this technically could
not be possible users are experiencing errors with this. I have added a line
of defensive programming to counter this issue.

In future audits of this part of the code we should ensure that this situation
cannot happen.