Simplified Twig writer
The twig writer was a complicated part of the codebase; as such it was hard to
test and it contained outdated concepts. One of these concepts was that the
router queue needs to be passed along; currently you can use the router queue
injected by the service container.

Although no additional test were written yet, the next step is to add tests for
the writer, the queue and the EnvironmentFactory