Documented Project Management and removed QA process
During SymfonyCon 2019 Jaap and I met and discussed how to work with
Github, and specifically projects. I decided to write a draft on how we
want to do project management for phpDocumentor so that it is clear for
new contributors, ourselves and possibly our users how we deal with
incoming requests, issues but also our roadmap.

I would like to stress that this is a draft that I would like to have
reviewed by the other team members.

As a sidenote, I also deleted the QA document from the internals folder
because it was heavily outdated. A new document needs to be written in
its place but that it out of scope for this change.