PHP needs to be invoked when running the PHAR
In the build PHP must be invoked to start the PHAR. This is because the
PHAR is not executable when downloaded and using chmod did not do the
trick. In addition, Windows does not have chmod, so it is more
cross-platform to actually call PHP and pass the PHAR as an argument.