Made LinkRenderer more readable
In this change I have made an effort to make the LinkRenderer more clear
in terms of readability and what it does. As far as I am concerned it is
still a monster of a class but I would like to capture its essence in
integration tests first. Hoping that that adds clarity and makes it
easier to pick apart.

Also: this change will bring coverage to 60%!