Introduce Symfony Config for configuration parsing
In this change I have introduced a first version of a Symfony
configuration definition that can parse the XML for Version 3 and
normalize that into an array and so validation on it. There are debug
statements in the code and I am will be continue building on this in
this POC branch.

I intend on squashing this into one commit with a new message :)