Add convenience binaries for tools
In order to make it easier to target specific files to test or to
provide formatters (for example for Behat) I have introduced a couple of
helper binaries in the bin folder.

These binaries are for:

- phpunit
- behat
- phpcs
- phpcbf

You can use these as if they were natively installed but nuder the hood
they call make and set the ARGS environment variable. This ARGS variable
is appended to the invocation of each tool and that way you can, for
example, do the following:

    ./bin/phpunit tests/unit/phpDocumentor/AutoloaderTest.php

or

    ./bin/behat --formar=progress

This can help in rerunning tests on specific classes or features that
you are working on, or using phpcbf to beautify one specific file.