Use global /tmp folder for cache pools
The cache pools now write into the cache folder for Symfony; however
this needs to be immutable for PHAR files to work correctly. As a first
experiment I have changed the location of these cache pools to /tmp;
however I am fairly confident that Windows installations won't like
this.

I want to check the impact on the build; and in a subsequent commit
improve on this so that the configurable cache folder may be used or at
least a folder that windows likes.