Add a smoke test stage
Since, for a yet unknown reason, Behat can't successfully execute the
PHAR with its tests (while locally this is not a problem), I have added
a smoke test that does nothing more than run the PHAR against
phpDocumentor itself.

This will help get more intel on why the behat builds fail; and is a
good check in case the behat tests miss an issue.