Replace Cocur/Slugify with Symfony/String
With phpDocumentor we are frequently running into transliteration and
slugification issues. For the slugify side of things we used
Cocur/Slugify but recently the Symfony/String component became available
that has support for various string manipulations, including
sluggification, that have transliteration of strings in mind.

With this change I hope to make a start to make the transliteration support
more robust for phpDocumentor and enable us to do more with this
feature.