Remove Parse and TransformCommand
In an effort to simplify the application I decided to remove the parse
and transform commands. In the past these were more meaningful because
the parse command created an XML file that could be read by other
applications; but since this is no longer the case there is less use in
maintaining these separate commands.