Add first test for Twig writer
The Twig writer is a rather complicated piece of work, in this first
test I have tackled the simplest scenario: rendering a twig file to a
single output file.

In subsequent commits this will be elaborated on by testing other
scenario's as well; such as generating a file for each class given a
specific template.