Extracted Destination determination into a service
The Twig writer uses various logical steps to determine where the output
of a twig writer action should be written to. This piece of code was
located in the WriterAbstract class but was untestable due to this.

By moving it to a service, it is now testable (tests not included) and
it helped simplify the WriterAbstract and Twig Writer because several
conditions and optional behaviour could be removed.