Move side-effects out of PathGenerator
The PathGenerator created an absolute path and created any missing
directories. By having the PathGenerator generate an absolute URL we can
centralize file system interactions to the Twig Writer (making it easier
to move to FlySystem) and simplify the EnvironmentFactory since it
actually makes a relative location out of an absolute one.