Add support for extensions
For phpDocumentor, we want to allow consumers and internal processes to
be able to manipulate what data will be rendered and with it the ability
to change what data is included.

Examples of this is the `@mixin` tag, with an extension you can read the
tags of all classes, determine whether it has a mixin tag and try to
locate an accompanying class that needs to be mixed in. Once found, you
can change the list of methods to include the mixed in methods; their
parent property remains a method to determine that they do not originate
in this class but instead from elsewhere.

This extension mechanism should also work when called from phpDocumentor
as a PHAR. Because you can register additional autoloading mechanisms
on runtime, it is able to query the real filesystem instead of inside
the PHAR. This means that projects could include their own extensions
and still refer to them from within the PHAR.

If this mechanism sticks, we could even attempt to move internal
processes (such as inheritence) to these extensions.