phpDocumentor goes into infinite loop when class extends itself
In the Symfony test suite is a test class that extends itself for
demonstration purposes (/src/Symfony/Component/VarDumper/Tests/Fixtures/NotLoadableClass.php),
this trips up phpDocumentor by going into an infinite loop.

This change will prevent setting the parent of a Descriptor when it
detects that the current FQSEN is the same as its parent FQSEN. This
will counter the exact example above.

This will not permanently solve all cyclic dependency induced infinite
loops but will fix the 80%. I have created an issue to add more
sophisticated tracking of cyclic dependencies.