FQSEN objects did not link
The linker converts (arrays of) FQSENs to Descriptors and injects these
into other Descriptors. This is used to restore the object graph after
serialization (object graphs cannot be serialized, only trees).

With an earlier refactoring the FQSEN object was introduced but the
linker was not updated to resolve this type; meaning that anything that
used an FQSEN did not resolve.