Remove Type descriptors
During the refactoring of the Linker I found that the Type descriptors
had been replaced by the reflection types. This commit contains the code
necessary to remove the types from the descriptors.

During a trial run of behat I found a couple of failures now, I will fix
those next