Add Search and a lot of little features to the new template
First I need to admit I should have split this commit, now it is a big
mess of all kinds of improvements. Sorry!

In this commit I have been working hard on the new template and fixing
small bugs that I encounter, such as missing links, too many links and
more.

The BIG bang for this change is the search functionality in the new
template. I have finally been able to figure out a way to have a good
working search AND support for offline files to use that search.