Add initial e2e test for Default template
Because Behat is not able to comfortably test frontends, I have done an
experiment using Cypress as a test runner. This change is the result of
this experiment.

In this change I have introduced a new make command `make test-template`
that will install Cypress using NPM, run phpDocumentor against the
Mario's Pizzeria example, and test the output using Cypress.

I have chosen not to include this in the build pipeline yet as this is
an example meant for feedback; if we continue this way I want to clean
up the make commands a bit more and add it to Github Actions.