Flesh out new Class Details page
In this change I have added an indicator in which file a method resides,
support for variadics, added BEM classes on everything for a class and
also added BEM classes that can be used to change styling when a method
is static, abstract, final and based on its visibility.

In addition, I have added various Cypress tests to verify at least most
of this behaviour.