Building an '@package' tree failed with certain characters
The `@package` tag supports various special characters historically that
FQSEN's do not. We do not want to maintain such complicated structures
next to how FQSEN's work; we can however normalize the PackageTree using
the following rules:

The characters '-', '.', '_', '[' and ']' in an `@package` are converted
into a `\` to match the FQSEN format and the output will have trailing
backslashes removed.

By removing trailing backslashes you prevent dangling separators, as you
would get with the example `\My[Space]`