Writers should be able to expose custom settings
It should be easy for services, among which writers, to be able to
expose (or even set) settings specific to them.

An example is included in this change where the graphs are only
generated if a setting called graphs.enabled is set to true.

Not included in this change is the ability to manipulate these settings
and to list them on the command line. This is planned for the next
commit