Custom Settings may be set from command line or config
In this change I have provided the ability to set custom settings from
the command line using `--setting KEY=VALUE` or `-s KEY=VALUE`. This
will trickle down to the settings object and invalidate the cache when
the settings have changed.

In addition, this also provides an option `--list-settings`. This option
will print a list of available settings; gathered from all services
implementing the WithCustomConfig interface. At this point this does not
yet print descriptions; this may be added in a future change.

Lastly, a new configuration option in the v3 options was added with
which you can define these settings using the config file.