Rework Guides Proof of Concept pt 1
The initial commit of the experimental Guides implementation was
barebones, did not make use of much of phpDocumentor's files and use of
FlySystem.

With this change, I have laid the groundworks for using phpDocumentor's
configuration system, together with FlySystem and more.

In order to verify this, and because it had to be done, I converted
phpDocumentor's config format to that of v3.