Render the 'default' template as part of a Guide
Prior to this change a guide did not include the header, styling and
other styling that the 'default' template has to offer. With this
change, I have been able to make that happen.

Unfortunately, this did impact the generation of the guides as we had to
ensure that the generation of guides was part of the transformation
stage instead of the parsing phase.

The Doctrine RST parser does not differentiate between parsing and
transforming; so I moved everything to the transformation stage.