Correctly render links to Nullable types
When a reference to an object is written as `?MyClass` it will generate
as a Nullable type. This was not supported by the LinkRenderer yet and
it will convert any Nullable into an array with the FQSEN and null; and
then pass it through the renderer again. This will ensure that the same
business logic is applied as when you use `MyClass|null`.