Allow use of 'local' templates
When a source location is not preceeded by 'templates', then it is
assumed to be used from inside the PHAR. Of course, this may cause
conflicts with users who have a templates folder that they use. This is
a more complex issue to solve and I would do that in a separate issue.