Reintroduce re-encoding of files
In issue #2254 it was reported that phpDocumentor was no longer able to
deal with ISO-8859-1 characters in a file.

After research, it was shown that during the migration from local files
to FlySystem the mechanism to re-encode files was not re-introduced.
This clearly causes issues and in this change I re-introduce this
behaviour as a Parser middleware.

This middleware will extract the file contents from FlySystem,
determines whether re-encoding is needed and does so. After this, a new
type of File is passed to the rest of the chain for processing.