Fixes #2277: Wrong ProjectDir is detected
When phpDocumentor is installed using Composer, the project directory as
detected by Symfony is in the wrong folder. Instead of picking the
actual root, it picks the `src` folder underneath.

By overriding the getProjectDir method, the algorithm used in
http-kernel's Kernel class does function correctly and the issue no
longer arises.

Don't. Ask. Why. Or. How. It works.