Check version on Run
With this change, phpDocumentor will fetch the releases from github and
check if the current version is the latest. If it is not, a message will
be shown to ask the user to upgrade.

In this change, I have also refactored away the VERSION file as we now
use PrettyVersions; and I have refactored the Application class away
since it did not contain anything useful or that should actually have
been elsewhere.