Add Files index and make more extendible
In this change I have introduced a new page in the default template that
will show all files in the project in alphabetical order.

In addition, I have added a level of indirection by moving the contents of
base.html.twig to layout.html.twig. This will enable people to override
base.html.twig in their .phpdoc folder; and to redefine specific blocks
instead of having to redefine the whole layout.