Ensure that phpDocumentor2 Wildcards are expanded
phpDocumentor 2 ignore statements in the configuration were not true
glob patterns and when you used a wildcard it literally meant 'any
number of characters in between'. This is incompatible with the way glob
patterns work and how some tools use wildcards. It also prevents more
fine-grained patterns as you cannot exclude the contents of a folder but
do include the contents of the subfolders underneath.

In phpDocumentor 3 we fixed this. This did introduce a backwards
incompatible change. To prevent all phpDocumentor 2 based config files I
have introduced an upgrade path in the phpDocumentor 2 config that any
ignore statement that ends with `/*` and is not `**/*` is interpreted as
if it were `**/*`.

This will allow us to consistently use glob patterns in the application
code, and only have the interpretation of the config worry about this
incompatibility.