Disable caching of twig templates and always enable debug
When developing on the templates it is quite confusing when the caches
are cleared or not. From a cursory investigation I can conclude that the
caches do not yield much of a benefit for us. The compilation of the
templates takes relatively little time and happens once at the beginning
of the application and not in the execution loop.

I have explicitly made this a separate commit, if I am proven wrong we
can revert this easily