Redesign header
Based on feedback and ongoing insights, we have made a new design that
will be easier to extend and fit better with other sites' designs so
that overrides can be smaller and easier.

This commit includes an experiment in splitting up parts of the template
into components, and to split the CSS off into component css files as
well.

Theoretically, this would make it easier to extend the styling as well
since you can override a specific component's CSS file instead of the
whole CSS or by using an override at the end that counters the CSS for
that component.