CI: allow Cypress tests to pass
This:
* Fixes one test - class-details / "Show methods with return type in the Table of Contents"
* Skips four tests which are currently completely broken.
* Partially skips two test which are currently partially broken.

The (partially) skipped tests examine the breadcrumbs and look for a `Home` link which doesn't exist.
These tests should either be removed or the behaviour should be fixed.

I'm kind of assuming this removal of the "Home" breadcrumb is intentional and the tests should be removed, but I didn't want to do so without verifying this. Please let me know if I can update the PR to remove these tests.