Docs: improvements for "references - phpdoc - tags - param"
Summary
* Added a link to the PHP manual page covering function arguments.

Description:
* Removed a redundant phrase - "Exceptions to this recommendation are:". Might have been orphaned after a previous edit ?

Example:
* Added a second argument with a different Type and a multi-line description.

Other:
* Verified the text with PSR 19.
* Minor inline markup and grammar fixes.
* Made the syntax outline a code block.
* Linked some text to related other documentation pages.

Ref: https://github.com/php-fig/fig-standards/blob/master/proposed/phpdoc-tags.md#59-param

**Open question**:

The specs feel outdated. If the function declaration has parameter type declarations, those should be leading and the tags can be omitted.
The specs currently do not explicitly allow for that.

In PSR-19, the following phrase has been added to the first paragraph of the description, which sort of allows for that.
Should this phrase be added or not ?

> The "name" is required only when some @param tags are omitted due to all useful info
> already being visible in the code signature itself.