Correctly render links to FQSEN references
See tags have a 'reference' to a destination. This reference can be an
FQSEN but also a URL. In case it was an FQSEN, the LinkRenderer still
interpreted this as a URL (as this is the default).

With this change, I have made the LinkRenderer aware of this type of
Link, and cast it to string so that the next check will automatically
transform it into an FQSEN.