Constant values in TOC should be singleline
Large constants may eat up the whole TOC. And while the values of
enum-style constants are convenient to see at a glance; these large
constants can really break the TOC.

In issue #2443, Juliette suggests to show the type instead. In my
opinion, this will diminish the value as you cannot see enums at a
glance and usually constants are simple primitives.

As an alternative solution to address this, I have altered the styling
of the TOC to cut off large values and show an ellipsis at the end of
the line.