Publish NPM package with assets
As we want to re-use the generated assets in the website and demo site,
I have created an NPM package based on the base assets from the default
template. The template has been designed in such a way that these other
sites should also be able to re-use the same assets.