[GUIDES] Use FlySystem for writing output to
The new guides system is based on Doctrine's RST Parser; that version
used direct writing and reading to and from disk. Because we are moving
towards a platform where it need not matter where your code is; I have
begun replacing the filesystem interaction with flysystem so that we can
connect phpDocumentor more easily onto this piece.