Admonition contents are not rendered
In this change, I have made a start with rendering admonitions. The
prior method did not include the body of the admonition and as such
didn't really show anything.

This can still be improved, both the docs and the rendering style; and I
will do so in a subsequent commit