Split transformer from template creation
To be able to add more transformers in an upcomming pr which will
allow us to support multiple versions, documentationsets etc. I removed
all template related code from the transformer.

The Template Factory will now provide a collection of templates to render
equal to the existing template collection this is able to provide the required
transformations. The Transformer will only instruct the writers to execute transformations.

This commit contains a number of shortcuts to finish this faster. The configuration type
hints could be improved, and the destination should be removed from the transformer. Although
both are not really required at the moment. Things will work as before now.