Render UML diagrams using plantuml
A new directive, `uml`, is introduced with which you can render uml
diagrams purely through text using the tool `plantuml`. For this to
work, a new directive needed to be added (jeez, need some rework here)
and a plantuml rendering service is introduced.

The presented code is functional, but can use some TLC and tests. For
one, I want to replace the innards of PlantumlClassDiagram to use the
new Renderer service.