Move init of twig to init step
By moving the init of twig to an init step we are clearing the road
to start processing documentation sets. The phpDoc twig extension uses the
context now to configure the router on render. Which removes direct calls to the
extension. And make it possible to switch context without touching the extension state.