Guides: Refactored NodeFactory and Instantiators away
By removing the environment and renderer services from the Node
instances; they can now be instantiated from any location in the code
and no longer rely on the instantiation piece of the process to form
them.

This meant that I could refactor the NodeFactory and instantiators away,
which makes it easier to add new directives and reduces the amount of
code to maintain and bring under test.