Guides: Remove createSpanNode from Parser
The Parser had a method createSpanNode that did nothing more than
instantiate a SpanNode; this code could easily be moved to the invoking
locations as this does not fall within the parser's responsibilities and
only incurs dependencies.

Although the dependencies are not directly solved with this change, it
does clean up the interface and paves the way for more improvements.