Guides: remove old classes and name more inline with the rest
This is one of the first steps into bringing Guides under test and
inline with the rest of phpDocumentor. By renaming the 'Scanner' to
FileCollector and by moving and removing code the whole should match the
naming and structure of phpDocumentor better.

This change also removes a couple of classes that were dead code.